On a way to save an int :D

+++ Assets/Scripts/PlayerProfile/Player.cs
@@ -1,6 +1,6 @@
 ﻿public class Player {
 
-    public int TotalSessions { get; set; }
+    public short TotalSessions { get; set; }
 
     public void IncrementSessionsCounter() {
         TotalSessions++;
